+++ dwertheimer.EventAutomations/src/presets.js
@@ -1,4 +1,4 @@
-import {logDebug} from '@helpers/dev'
+import { logDebug } from '@helpers/dev'
 
 function createConfigOptions() {}