+++ np.WeatherLookup/__tests__/utils.test.js
@@ -6,19 +6,19 @@ import * as w from '../src/support/weather-utils'
 describe('np.WeatherLookup' /* pluginID */, () => {
   describe('utils' /* file */, () => {
     describe('isWeatherKeyValid' /* function */, () => {
-      test('should return false on empty string',  () => {
+      test('should return false on empty string', () => {
         const result = w.isWeatherKeyValid('')
         expect(result).toEqual(false)
       })
-      test('should return false on undefined string',  () => {
+      test('should return false on undefined string', () => {
         const result = w.isWeatherKeyValid(undefined)
         expect(result).toEqual(false)
       })
-      test('should return false on wrong string',  () => {
+      test('should return false on wrong string', () => {
         const result = w.isWeatherKeyValid(`foo`)
         expect(result).toEqual(false)
       })
-      test('should return true on correct string signature',  () => {
+      test('should return true on correct string signature', () => {
         const result = w.isWeatherKeyValid(`11634c5bc8f3ac1aa1442085146b969a`)
         expect(result).toEqual(true)
       })